- Replace use of memcpy for write and read pixel macros because float4 is neither plain data nor packed. Anyways a few conditionals with assignments is better optimized by compiler. - Fix scaling operations issue. Float coordinates were being casted to int even for bilinear sampling. - Fix saved files being not compatible with previous versions due to removal of tree chunksize field. chunksize field is added again just for compatibility.

+++ source/blender/compositor/buffering/COM_Rect.cpp
@@ -90,8 +90,8 @@ PixelsImg PixelsRect::pixelsImgCustom(
   int elem_chs = n_channels;
   size_t elem_bytes = elem_chs * sizeof(float);
 
-  int brow_elems = buffer_row_bytes / elem_bytes;
-  int brow_chs = brow_elems * elem_chs;
+  size_t brow_elems = buffer_row_bytes / elem_bytes;
+  size_t brow_chs = brow_elems * elem_chs;
 
   float *buffer_y_start = is_single_elem ? buffer :
                                            buffer + (size_t)rect.ymin * brow_elems * n_channels;
